Broke up util into separate packages
